Deleted EditUserForm and modified CreateUserForm to have attributes interests and location for Edit user.
