+++ src/main/java/com/newts/newtapp/api/forms/CreateUserForm.java
@@ -1,5 +1,7 @@
 package com.newts.newtapp.api.forms;
 
+import java.util.ArrayList;
+
 /**
  * A form defining how a create user request's json request body should be formatted for Spring to serialize.
  */
@@ -7,11 +9,15 @@ public class CreateUserForm {
     private final String username;
     private final String password;
     private final String interest;
+    private final ArrayList<String> interests;
+    private final String location;
 
-    public CreateUserForm(String username, String password, String interest) {
+    public CreateUserForm(String username, String password, String interest, ArrayList<String> interests, String location) {
         this.username = username;
         this.password = password;
         this.interest = interest;
+        this.interests = interests;
+        this.location = location;
     }
 
     public String getUsername() {
@@ -24,5 +30,11 @@ public class CreateUserForm {
 
     public String getInterest() {
         return interest;
+    }
+    public ArrayList<String> getInterests() {
+        return interests;
+    }
+    public String getgetLocation() {
+        return location;
     }
 }
